@@ -1,14 +1,24 @@
 module ngssc
 
-go 1.17
+go 1.20
 
 require (
 	github.com/bmatcuk/doublestar v1.3.2
+	github.com/fsnotify/fsnotify v1.6.0
+	github.com/hashicorp/go-envparse v0.1.0
 	github.com/urfave/cli v1.22.4
+	golang.org/x/exp v0.0.0-20230801115018-d63ba01acd4b
 )
 
 require (
-	github.com/cpuguy83/go-md2man/v2 v2.0.0-20190314233015-f79a8a8ca69d // indirect
-	github.com/russross/blackfriday/v2 v2.0.1 // indirect
-	github.com/shurcooL/sanitized_anchor_name v1.0.0 // indirect
+	github.com/xrash/smetrics v0.0.0-20201216005158-039620a65673 // indirect
+	golang.org/x/sys v0.1.0 // indirect
+)
+
+require (
+	github.com/andybalholm/brotli v1.0.5
+	github.com/cpuguy83/go-md2man/v2 v2.0.2 // indirect
+	github.com/hashicorp/golang-lru/v2 v2.0.2
+	github.com/russross/blackfriday/v2 v2.1.0 // indirect
+	github.com/urfave/cli/v2 v2.25.3
 )
