@@ -1,4 +1,4 @@
-package main
+package insert
 
 import (
 	"fmt"
