@@ -1,7 +1,14 @@
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import { bootstrapApplication } from '@angular/platform-browser';
+import { CSP_NONCE } from '@angular/core';
+import { environment } from './environments/environment';
+import { TITLE_TOKEN, AppComponent } from './app/app.component';
 
-import { AppModule } from './app/app.module';
-
-platformBrowserDynamic()
-  .bootstrapModule(AppModule)
-  .catch((err) => console.error(err));
+bootstrapApplication(AppComponent, {
+  providers: [
+    { provide: TITLE_TOKEN, useValue: environment.title },
+    {
+      provide: CSP_NONCE,
+      useValue: environment.cspNonce,
+    },
+  ],
+}).catch((err) => console.error(err));
